# perf(http): asynchronous retrieval of user keys