### Description

Speeds up request by more than 3-4x.

Before: `make run  0.44s user 0.70s system 19% cpu 5.727 total`

After: `make run  0.59s user 0.79s system 92% cpu 1.499 total`